gosu 폴더 생성 후 고수 관계 페이지를 gosu 폴더에 옮기고 오타 고침
