+++ zoomgofrontend/src/css/main/gosu_header.css
@@ -44,7 +44,7 @@ button {
   color: white;
   margin-left: 80px;
 }
-.gosu_banner > button:hover {
+.banner_button:hover {
   cursor: pointer;
   background-color: #6788ff;
 }
@@ -89,13 +89,6 @@ button {
   margin-left: 30px;
   color: black;
 }
-.gosu_experience {
-  margin-top: 50px;
-  text-align: right;
-  font-size: 10px;
-  color: rgba(0, 0, 0, 0.521);
-}
-
 .gosu_profileList button {
   width: 90px;
   border-radius: 10px;
