+++ zoomgofrontend/src/css/main/section3.css
@@ -46,6 +46,7 @@
 .reivew_top > span {
     font-size:14px;
     font-weight:bold;
+    
 }
 .review_work a{
     color:#c8c8c8;
