+++ zoomgofrontend/src/css/main/section4.css
@@ -58,3 +58,13 @@
     margin-top:70px;
     margin-left:10px;
 }
+
+.gosu_experience {
+    margin-top: 50px;
+    text-align: right;
+    font-size: 10px;
+    color: rgba(0, 0, 0, 0.521);
+  }
+  .move_gosuporfolio_list {
+    border:0;
+  }
