+++ zoomgofrontend/src/forgotpassword/forgotpassword.css
@@ -1,6 +1,3 @@
-/* body {
-    font-family: Arial, sans-serif;
-} */
 
 .forgotpassword1 {
     margin: 20px;
