+++ zoomgofrontend/src/gosu/category_portfolio.js
@@ -1,4 +1,4 @@
-import "./css/portfolio/portfolio_category.css";
+import "../css/portfolio/portfolio_category.css";
 import Header from "./gosu_header";
 function CategoryPortfolio() {
   const categoryList = ["가", "나", "다", "라", "마", "바"];
