+++ zoomgofrontend/src/gosu/portfolio.js
@@ -1,8 +1,8 @@
 import { FaAngleLeft, FaAngleRight, FaStar } from "react-icons/fa";
 import { FaRegShareFromSquare } from "react-icons/fa6";
-import "./css/portfolio/portfolio.css";
+import "../css/portfolio/portfolio.css";
+import dogs from "../img/dog.png";
 import Header from "./gosu_header";
-import dogs from "./img/dog.png";
 function Portfolio() {
   return (
     <div className="App">
