+++ zoomgofrontend/src/gosu/profile.input.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import "./css/profile/career.css";
-import "./css/profile/profileInput.css";
+import "../css/profile/career.css";
+import "../css/profile/profileInput.css";
 import Header from "./gosu_header";
 function ProfileInput() {
   const [fileName, setFileName] = useState("");
