+++ zoomgofrontend/src/gosu/profile.js
@@ -1,5 +1,5 @@
-import './css/profile/middle.css';
-import './css/profile/top.css';
+import '../css/profile/middle.css';
+import '../css/profile/top.css';
 import Header from './gosu_header';
 
 import { FaStar } from "react-icons/fa";
