+++ zoomgofrontend/src/registpage/Registpage.css
@@ -1,9 +1,4 @@
-/* body {
-    background-color: #ffffff;
-    display: flex; 
-    justify-content: center; 
-    align-items: center; 
-} */
+
   
 .container {
     padding: 32px;
@@ -48,6 +43,7 @@ input:-ms-input-placeholder{
     width: 230px;
     padding: 17px;
     border-radius: 20px; 
+    border:none;
 }
 
 .gogodeta {
@@ -73,9 +69,6 @@ input:-ms-input-placeholder{
     cursor: pointer;
 }
   
-button:hover {
-    background-color: #6788FF;
-}
 
 .regist-main{
     margin-bottom: 80px;
